Added code for square root
The function takes 2 arguments:
x: The number that we want to calculate the square root of.
x0: An initial guess as to what the square root is. Is it possible to initially send (x / 2.0) as the parameter for x0?

The function recursively calls itself until the difference between iteration n and n-1 is less than epsilon.